+++ python/lvmecp/actor/commands/status.py
@@ -46,4 +46,6 @@ async def status(command: ECPCommand, no_registers: bool = False):
         o2_percent_spectrograph=plc.safety.o2_level_spectrograph,
     )
 
+    command.info(last_heartbeat_set=command.actor._last_heartbeat)
+
     return command.finish()